+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Migrated `ScanSimpleIdScreen` to Mobius
 - Migrate `EnterOtpScreen` to Mobius
 - Setup the UI via an effect in `PatientEntryScreen`
+- Setup the UI via an effect in `EditPatientScreen`
 - [In Progress: 06 Aug 2020] Migrate `PatientSearchResultsScreen` to Mobius
 - [In Progress: 10 Aug 2020] Migrate `ConfirmRemoveBloodPressureDialog` to Mobius